New inspection rule for localextensions.xml - Unknown Extension declared as dependency
