+++ src/com/intellij/idea/plugin/hybris/codeInspection/rule/extensioninfo/AbstractEiInspection.kt
@@ -16,21 +16,16 @@
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package com.intellij.idea.plugin.hybris.system.extensionInfo
+package com.intellij.idea.plugin.hybris.codeInspection.rule.extensioninfo
 
-import com.intellij.idea.plugin.hybris.common.HybrisConstants
-import com.intellij.idea.plugin.hybris.common.utils.HybrisIcons
-import com.intellij.idea.plugin.hybris.system.extensionInfo.model.ExtensionInfo
-import com.intellij.openapi.module.Module
+import com.intellij.idea.plugin.hybris.codeInspection.rule.AbstractInspection
+import com.intellij.idea.plugin.hybris.system.extensioninfo.model.ExtensionInfo
+import com.intellij.openapi.project.Project
 import com.intellij.psi.xml.XmlFile
-import com.intellij.util.xml.DomFileDescription
-import javax.swing.Icon
 
-class EiSDomFileDescription : DomFileDescription<ExtensionInfo>(ExtensionInfo::class.java, "extensioninfo") {
+abstract class AbstractEiInspection : AbstractInspection<ExtensionInfo>(ExtensionInfo::class.java) {
 
-    override fun getFileIcon(flags: Int): Icon = HybrisIcons.EXTENSION_INFO
+    override fun canProcess(project: Project, file: XmlFile) = true
 
-    override fun isMyFile(file: XmlFile, module: Module?) = super.isMyFile(file, module)
-            && file.name == HybrisConstants.EXTENSION_INFO_XML
-
+    override fun canProcess(dom: ExtensionInfo) = dom.xmlElement != null
 }
