+++ src/com/intellij/idea/plugin/hybris/system/extensioninfo/EiSDomFileDescription.kt
@@ -16,33 +16,21 @@
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
 
-// Generated on Mon Jan 09 16:31:21 CET 2023
-// DTD/Schema  :    null
-package com.intellij.idea.plugin.hybris.system.extensionInfo.model;
+package com.intellij.idea.plugin.hybris.system.extensioninfo
 
-import com.intellij.util.xml.*;
-import com.intellij.util.xml.DomElement;
-import org.jetbrains.annotations.NotNull;
+import com.intellij.idea.plugin.hybris.common.HybrisConstants
+import com.intellij.idea.plugin.hybris.common.utils.HybrisIcons
+import com.intellij.idea.plugin.hybris.system.extensioninfo.model.ExtensionInfo
+import com.intellij.openapi.module.Module
+import com.intellij.psi.xml.XmlFile
+import com.intellij.util.xml.DomFileDescription
+import javax.swing.Icon
 
-/**
- * null:extensioninfoElemType interface.
- * <pre>
- * <h3>Type null:extensioninfoElemType documentation</h3>
- * Configures the available modules of the extension.
- * </pre>
- */
-@Stubbed
-@StubbedOccurrence
-public interface ExtensionInfo extends DomElement {
+class EiSDomFileDescription : DomFileDescription<ExtensionInfo>(ExtensionInfo::class.java, "extensioninfo") {
 
-	/**
-	 * Returns the value of the extension child.
-	 * @return the value of the extension child.
-	 */
-	@NotNull
-	@SubTag ("extension")
-	@Required
-	Extension getExtension();
+    override fun getFileIcon(flags: Int): Icon = HybrisIcons.EXTENSION_INFO
 
+    override fun isMyFile(file: XmlFile, module: Module?) = super.isMyFile(file, module)
+            && file.name == HybrisConstants.EXTENSION_INFO_XML
 
 }
