+++ src/com/intellij/idea/plugin/hybris/system/extensioninfo/model/WebModule.java
@@ -18,20 +18,34 @@
 
 // Generated on Mon Jan 09 16:31:21 CET 2023
 // DTD/Schema  :    null
-package com.intellij.idea.plugin.hybris.system.extensionInfo.model;
+package com.intellij.idea.plugin.hybris.system.extensioninfo.model;
 
 import com.intellij.util.xml.*;
 import com.intellij.util.xml.DomElement;
 import org.jetbrains.annotations.NotNull;
 
 /**
- * null:hmcmoduleType interface.
+ * null:webmoduleType interface.
  * <pre>
- * <h3>Type null:hmcmoduleType documentation</h3>
- * Configures an hmc module for the extension. Required directory: /hmc.
+ * <h3>Type null:webmoduleType documentation</h3>
+ * Configures an hMC module for the extension. Required directory: /web.
  * </pre>
  */
-public interface HmcModule extends DomElement {
+public interface WebModule extends DomElement {
+
+	/**
+	 * Returns the value of the webroot child.
+	 * <pre>
+	 * <h3>Attribute null:webroot documentation</h3>
+	 * Webroot where the web application will be available at.
+	 * </pre>
+	 * @return the value of the webroot child.
+	 */
+	@NotNull
+	@Attribute ("webroot")
+	@Required
+	GenericAttributeValue<String> getWebroot();
+
 
 	/**
 	 * Returns the value of the additionalclasspath child.
@@ -47,23 +61,23 @@ public interface HmcModule extends DomElement {
 
 
 	/**
-	 * Returns the value of the extensionclassname child.
+	 * Returns the value of the jspcompile child.
 	 * <pre>
-	 * <h3>Attribute null:extensionclassname documentation</h3>
-	 * Name of the extension's HMCExtension class.
+	 * <h3>Attribute null:jspcompile documentation</h3>
+	 * If "true", JSP files will be pre-compiled as part of the build process. If "false", JSP files will be compiled when first used by the application server. Default is "true".
 	 * </pre>
-	 * @return the value of the extensionclassname child.
+	 * @return the value of the jspcompile child.
 	 */
 	@NotNull
-	@Attribute ("extensionclassname")
-	GenericAttributeValue<String> getExtensionClassname();
+	@Attribute ("jspcompile")
+	GenericAttributeValue<Boolean> getJspCompile();
 
 
 	/**
 	 * Returns the value of the sourceavailable child.
 	 * <pre>
 	 * <h3>Attribute null:sourceavailable documentation</h3>
-	 * Deprecated. Has no effect and will be evaluated always to 'true' if a 'hmc/src' directory is available
+	 * Deprecated. Has no effect and will be evaluated always to 'true' if a 'web/src' directory is available
 	 * </pre>
 	 * @return the value of the sourceavailable child.
 	 */
